backend/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
curl_cffi==0.13.0
distro==1.9.0
fastapi==0.120.3
frozendict==2.4.6
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.0.3
langchain-core==1.0.2
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.38
lxml==6.0.2
multitasking==0.0.12
numpy==2.3.4
openai==2.6.1
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
peewee==3.18.3
platformdirs==4.5.0
portalocker==3.2.0
protobuf==6.33.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.1
-e git+ssh://git@github.com/pr-suresh/stock-research-assistant-v0.git@bd7940d76eaa36f1910c9b849bbbeb9d87135e9f#egg=stock_research_assistant
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yfinance==0.2.66
zstandard==0.25.0
